Tidy indentation of subsequent lines in media queries.
In the status quo, if there is a multi-tag selector inside of a media
query, and the `$tagSeparator` includes a newline, the subsequent lines
aren’t indented as they should be.

Now, we clearly don’t want to just shove `$pre` in between every
selector if the `$tagSeparator` doesn’t contain a newline.

So let’s check for it, and switch to a local version of
`$this->tagSeparator` so it doesn’t mess with the full property — and
if there is a new line in the `$tagSeparator`, append it with the
`$pre` indent.