Fix for "continue" causing a warning in PHP 7.3
"continue statements targeting switch control flow structures will now generate a warning. In PHP such continue statements are equivalent to break, while they behave as continue 2 in other languages."
http://php.net/manual/en/migration73.incompatible.php

"In PHP the switch statement is considered a looping structure for the purposes of continue. continue behaves like break (when no arguments are passed). If a switch is inside a loop, continue 2 will continue with the next iteration of the outer loop."
http://php.net/manual/en/control-structures.continue.php

Signed-off-by: Vince <git@darkain.com>